AG-14268 - improved code to refocus rows that were removed
