+++ packages/ag-grid-community/src/navigation/cellNavigationService.ts
@@ -3,12 +3,10 @@ import type { NamedBean } from '../context/bean';
 import { BeanStub } from '../context/beanStub';
 import type { BeanCollection } from '../context/context';
 import type { AgColumn } from '../entities/agColumn';
+import { _getRowAbove, _getRowBelow } from '../entities/positionUtils';
 import type { RowNode } from '../entities/rowNode';
-import { _isGroupRowsSticky } from '../gridOptionsUtils';
 import type { CellPosition } from '../interfaces/iCellPosition';
 import type { IRowNode } from '../interfaces/iRowNode';
-import type { RowPosition } from '../interfaces/iRowPosition';
-import type { RowCtrl } from '../rendering/row/rowCtrl';
 import type { RowSpanService } from '../rendering/spanning/rowSpanService';
 import { _last } from '../utils/array';
 import { _missing } from '../utils/generic';
@@ -171,82 +169,6 @@ export class CellNavigationService extends BeanStub implements NamedBean {
         } as CellPosition;
     }
 
-    public getRowBelow(rowPosition: RowPosition): RowPosition | null {
-        // if already on top row, do nothing
-        const index = rowPosition.rowIndex;
-        const pinned = rowPosition.rowPinned;
-        let ignoreSticky = false;
-        const { pageBounds, pinnedRowModel, rowModel } = this.beans;
-        if (this.isLastRowInContainer(rowPosition)) {
-            switch (pinned) {
-                case 'bottom':
-                    // never any rows after pinned bottom
-                    return null;
-                case 'top':
-                    // if on last row of pinned top, then next row is main body (if rows exist),
-                    // otherwise it's the pinned bottom
-                    if (rowModel.isRowsToRender()) {
-                        return { rowIndex: pageBounds.getFirstRow(), rowPinned: null } as RowPosition;
-                    }
-
-                    if (pinnedRowModel?.isRowsToRender('bottom')) {
-                        return { rowIndex: 0, rowPinned: 'bottom' } as RowPosition;
-                    }
-
-                    return null;
-                default:
-                    // if in the main body, then try pinned bottom, otherwise return nothing
-                    if (pinnedRowModel?.isRowsToRender('bottom')) {
-                        return { rowIndex: 0, rowPinned: 'bottom' } as RowPosition;
-                    }
-                    return null;
-            }
-        } else if (pinned) {
-            // if more pinned rows, should always navigate there
-            ignoreSticky = true;
-        }
-
-        const rowNode = rowModel.getRow(rowPosition.rowIndex);
-
-        const nextStickyPosition = ignoreSticky ? undefined : this.getNextStickyPosition(rowNode);
-
-        if (nextStickyPosition) {
-            return nextStickyPosition;
-        }
-
-        return { rowIndex: index + 1, rowPinned: pinned } as RowPosition;
-    }
-
-    private getNextStickyPosition(rowNode?: RowNode, up?: boolean): RowPosition | undefined {
-        const { gos, rowRenderer } = this.beans;
-        if (!_isGroupRowsSticky(gos) || !rowNode || !rowNode.sticky) {
-            return;
-        }
-
-        const isTopCtrls = rowRenderer
-            .getStickyTopRowCtrls()
-            .some((ctrl) => ctrl.rowNode.rowIndex === rowNode.rowIndex);
-
-        let stickyRowCtrls: RowCtrl[] = [];
-        if (isTopCtrls) {
-            stickyRowCtrls = [...rowRenderer.getStickyTopRowCtrls()].sort(
-                (a, b) => a.rowNode.rowIndex! - b.rowNode.rowIndex!
-            );
-        } else {
-            stickyRowCtrls = [...rowRenderer.getStickyBottomRowCtrls()].sort(
-                (a, b) => b.rowNode.rowIndex! - a.rowNode.rowIndex!
-            );
-        }
-
-        const diff = up ? -1 : 1;
-        const idx = stickyRowCtrls.findIndex((ctrl) => ctrl.rowNode.rowIndex === rowNode.rowIndex);
-        const nextCtrl = stickyRowCtrls[idx + diff];
-
-        if (nextCtrl) {
-            return { rowIndex: nextCtrl.rowNode.rowIndex!, rowPinned: null };
-        }
-    }
-
     private getCellBelow(lastCell: CellPosition | null): CellPosition | null {
         if (!lastCell) {
             return null;
@@ -255,7 +177,7 @@ export class CellNavigationService extends BeanStub implements NamedBean {
         // adjust spanned cell so when moving down asserts use of last row in cell
         const adjustedLastCell = this.rowSpanSvc?.getCellEnd(lastCell) ?? lastCell;
 
-        const rowBelow = this.getRowBelow(adjustedLastCell);
+        const rowBelow = _getRowBelow(this.beans, adjustedLastCell);
         if (rowBelow) {
             return {
                 rowIndex: rowBelow.rowIndex,
@@ -267,79 +189,6 @@ export class CellNavigationService extends BeanStub implements NamedBean {
         return null;
     }
 
-    private isLastRowInContainer(rowPosition: RowPosition): boolean {
-        const pinned = rowPosition.rowPinned;
-        const index = rowPosition.rowIndex;
-
-        const { pinnedRowModel, pageBounds } = this.beans;
-
-        if (pinned === 'top') {
-            const lastTopIndex = pinnedRowModel?.getPinnedTopRowCount() ?? 0 - 1;
-            return lastTopIndex <= index;
-        }
-
-        if (pinned === 'bottom') {
-            const lastBottomIndex = pinnedRowModel?.getPinnedBottomRowCount() ?? 0 - 1;
-            return lastBottomIndex <= index;
-        }
-
-        const lastBodyIndex = pageBounds.getLastRow();
-        return lastBodyIndex <= index;
-    }
-
-    public getRowAbove(rowPosition: RowPosition): RowPosition | null {
-        // if already on top row, do nothing
-        const index = rowPosition.rowIndex;
-        const pinned = rowPosition.rowPinned;
-        const { pageBounds, pinnedRowModel, rowModel } = this.beans;
-        const isFirstRow = pinned ? index === 0 : index === pageBounds.getFirstRow();
-        let ignoreSticky = false;
-
-        const getLastFloatingTopRow = (): RowPosition => {
-            const lastFloatingRow = pinnedRowModel?.getPinnedTopRowCount() ?? 0 - 1;
-
-            return { rowIndex: lastFloatingRow, rowPinned: 'top' } as RowPosition;
-        };
-
-        // if already on top row, do nothing
-        if (isFirstRow) {
-            if (pinned === 'top') {
-                return null;
-            }
-
-            if (!pinned) {
-                if (pinnedRowModel?.isRowsToRender('top')) {
-                    return getLastFloatingTopRow();
-                }
-                return null;
-            }
-
-            // last floating bottom
-            if (rowModel.isRowsToRender()) {
-                const lastBodyRow = pageBounds.getLastRow();
-                return { rowIndex: lastBodyRow, rowPinned: null } as RowPosition;
-            }
-
-            if (pinnedRowModel?.isRowsToRender('top')) {
-                return getLastFloatingTopRow();
-            }
-
-            return null;
-        } else if (pinned) {
-            // if more pinned rows, should always navigate there
-            ignoreSticky = true;
-        }
-
-        const rowNode = rowModel.getRow(rowPosition.rowIndex);
-        const nextStickyPosition = ignoreSticky ? undefined : this.getNextStickyPosition(rowNode, true);
-
-        if (nextStickyPosition) {
-            return nextStickyPosition;
-        }
-
-        return { rowIndex: index - 1, rowPinned: pinned } as RowPosition;
-    }
-
     private getCellAbove(lastCell: CellPosition | null): CellPosition | null {
         if (!lastCell) {
             return null;
@@ -348,7 +197,7 @@ export class CellNavigationService extends BeanStub implements NamedBean {
         // adjust spanned cell so when moving up asserts use of first row in cell
         const adjustedLastCell = this.rowSpanSvc?.getCellStart(lastCell) ?? lastCell;
 
-        const rowAbove = this.getRowAbove({
+        const rowAbove = _getRowAbove(this.beans, {
             rowIndex: adjustedLastCell.rowIndex,
             rowPinned: adjustedLastCell.rowPinned,
         });
@@ -386,7 +235,7 @@ export class CellNavigationService extends BeanStub implements NamedBean {
         if (!newColumn) {
             newColumn = displayedColumns[0];
 
-            const rowBelow = this.getRowBelow(gridCell);
+            const rowBelow = _getRowBelow(this.beans, gridCell);
             if (_missing(rowBelow)) {
                 return null;
             }
@@ -405,7 +254,8 @@ export class CellNavigationService extends BeanStub implements NamedBean {
     }
 
     public getNextTabbedCellBackwards(gridCell: CellPosition): CellPosition | null {
-        const { visibleCols, pagination } = this.beans;
+        const { beans } = this;
+        const { visibleCols, pagination } = beans;
         const displayedColumns = visibleCols.allCols;
 
         let newRowIndex: number | null = gridCell.rowIndex;
@@ -418,7 +268,7 @@ export class CellNavigationService extends BeanStub implements NamedBean {
         if (!newColumn) {
             newColumn = _last(displayedColumns);
 
-            const rowAbove = this.getRowAbove({ rowIndex: gridCell.rowIndex, rowPinned: gridCell.rowPinned });
+            const rowAbove = _getRowAbove(beans, { rowIndex: gridCell.rowIndex, rowPinned: gridCell.rowPinned });
 
             if (_missing(rowAbove)) {
                 return null;
